Build: Improve docs around config.json file
There are two versions of this file on different jq servers internally,
one with just the wp stuff, one with just the cdn stuff.

Update the sample file to match the keys that exist in reality.
Specifically, the sample was missing the 'zone_id' field that
we were already depending on in purge.php.
The sample was also missing the 'alias' field that we
set in production. Make use of this field instead of hardcoding
"jquery" so that things are easier to update quickly when needed.