Add redirects for git releases
This commit adds redirects for working versions of the code ("git") so that they can be kept on
releases.jquery.com instead of codeorigin. This is being done because Jenkins pushes these build
artifacts to the server automatically, and moving these files to releases means we don't have to
grant Jenkins access to codeorigin.

Signed-off-by: Brian Warner <brian@bdwarner.com>