qunit: Fix stylesheet link to not put up JS dialog
It would be nice to support `<link>` in this dialog, as requested
in https://github.com/jquery/codeorigin.jquery.com/issues/71, but right
now the dialog suggests `<script>` for the stylesheet which is wrong.

This only affected QUnit, in which I mistakenly wrote it this way
since commit 6e82c21c72 in 2018.

Ref https://github.com/jquery/codeorigin.jquery.com/pull/41.